ids de int a long

+++ RoutinesGymService.Application.DataTransferObject/Entity/SplitDayDTO.cs
@@ -5,7 +5,7 @@ namespace RoutinesGymService.Application.DataTransferObject.Entity
     public class SplitDayDTO
     {
         public WeekDay DayName { get; set; } 
-        public int RoutineId { get; set; }
+        public long RoutineId { get; set; }
         public string DayExercisesDescription { get; set; } = string.Empty;
         public List<ExerciseDTO> Exercises { get; set; } = new List<ExerciseDTO>();
     }
